+++ src/shell_app/launch_shell_app.js
